+++ stac_fastapi/api/stac_fastapi/api/app.py
@@ -1,7 +1,7 @@
 """Fastapi app creation."""
 
 
-from typing import Dict, List, Optional, Tuple, Type, Union
+from typing import Awaitable, Callable, Dict, List, Optional, Tuple, Type, Union
 
 import attr
 from brotli_asgi import BrotliMiddleware
@@ -25,7 +25,12 @@ from stac_fastapi.api.models import (
     ItemUri,
 )
 from stac_fastapi.api.openapi import update_openapi
-from stac_fastapi.api.routes import Scope, add_route_dependencies, create_async_endpoint
+from stac_fastapi.api.routes import (
+    Scope,
+    add_direct_response,
+    add_route_dependencies,
+    create_async_endpoint,
+)
 from stac_fastapi.types.config import ApiSettings, Settings
 from stac_fastapi.types.core import AsyncBaseCoreClient, BaseCoreClient
 from stac_fastapi.types.extension import ApiExtension
@@ -62,6 +67,10 @@ class StacApi:
             specified routes. This is useful
             for applying custom auth requirements to routes defined elsewhere in
             the application.
+        health_check:
+            A Callable which return application's `health` information.
+            Defaults to `def health: return {"status": "UP"}`
+
     """
 
     settings: ApiSettings = attr.ib()
@@ -123,6 +132,9 @@ class StacApi:
         )
     )
     route_dependencies: List[Tuple[List[Scope], List[Depends]]] = attr.ib(default=[])
+    health_check: Union[Callable[[], Dict], Callable[[], Awaitable[Dict]]] = attr.ib(
+        default=lambda: {"status": "UP"}
+    )
 
     def get_extension(self, extension: Type[ApiExtension]) -> Optional[ApiExtension]:
         """Get an extension.
@@ -138,12 +150,8 @@ class StacApi:
                 return ext
         return None
 
-    def register_landing_page(self):
-        """Register landing page (GET /).
-
-        Returns:
-            None
-        """
+    def register_landing_page(self) -> None:
+        """Register landing page (GET /)."""
         self.router.add_api_route(
             name="Landing Page",
             path="/",
@@ -165,12 +173,8 @@ class StacApi:
             endpoint=create_async_endpoint(self.client.landing_page, EmptyRequest),
         )
 
-    def register_conformance_classes(self):
-        """Register conformance classes (GET /conformance).
-
-        Returns:
-            None
-        """
+    def register_conformance_classes(self) -> None:
+        """Register conformance classes (GET /conformance)."""
         self.router.add_api_route(
             name="Conformance Classes",
             path="/conformance",
@@ -192,12 +196,8 @@ class StacApi:
             endpoint=create_async_endpoint(self.client.conformance, EmptyRequest),
         )
 
-    def register_get_item(self):
-        """Register get item endpoint (GET /collections/{collection_id}/items/{item_id}).
-
-        Returns:
-            None
-        """
+    def register_get_item(self) -> None:
+        """Register get item endpoint (GET /collections/{collection_id}/items/{item_id})."""  # noqa: E501
         self.router.add_api_route(
             name="Get Item",
             path="/collections/{collection_id}/items/{item_id}",
@@ -219,12 +219,8 @@ class StacApi:
             ),
         )
 
-    def register_post_search(self):
-        """Register search endpoint (POST /search).
-
-        Returns:
-            None
-        """
+    def register_post_search(self) -> None:
+        """Register search endpoint (POST /search)."""
         self.router.add_api_route(
             name="Search",
             path="/search",
@@ -248,12 +244,8 @@ class StacApi:
             ),
         )
 
-    def register_get_search(self):
-        """Register search endpoint (GET /search).
-
-        Returns:
-            None
-        """
+    def register_get_search(self) -> None:
+        """Register search endpoint (GET /search)."""
         self.router.add_api_route(
             name="Search",
             path="/search",
@@ -277,12 +269,8 @@ class StacApi:
             ),
         )
 
-    def register_get_collections(self):
-        """Register get collections endpoint (GET /collections).
-
-        Returns:
-            None
-        """
+    def register_get_collections(self) -> None:
+        """Register get collections endpoint (GET /collections)."""
         self.router.add_api_route(
             name="Get Collections",
             path="/collections",
@@ -306,12 +294,8 @@ class StacApi:
             ),
         )
 
-    def register_get_collection(self):
-        """Register get collection endpoint (GET /collection/{collection_id}).
-
-        Returns:
-            None
-        """
+    def register_get_collection(self) -> None:
+        """Register get collection endpoint (GET /collection/{collection_id})."""
         self.router.add_api_route(
             name="Get Collection",
             path="/collections/{collection_id}",
@@ -335,12 +319,8 @@ class StacApi:
             ),
         )
 
-    def register_get_item_collection(self):
-        """Register get item collection endpoint (GET /collection/{collection_id}/items).
-
-        Returns:
-            None
-        """
+    def register_get_item_collection(self) -> None:
+        """Register get item collection endpoint (GET /collection/{collection_id}/items)."""  # noqa: E501
         self.router.add_api_route(
             name="Get ItemCollection",
             path="/collections/{collection_id}/items",
@@ -364,7 +344,7 @@ class StacApi:
             ),
         )
 
-    def register_core(self):
+    def register_core(self) -> None:
         """Register core STAC endpoints.
 
             GET /
@@ -378,8 +358,6 @@ class StacApi:
 
         Injects application logic (StacApi.client) into the API layer.
 
-        Returns:
-            None
         """
         self.register_landing_page()
         self.register_conformance_classes()
@@ -390,41 +368,68 @@ class StacApi:
         self.register_get_collection()
         self.register_get_item_collection()
 
-    def add_health_check(self):
+    def add_health_check(self) -> None:
         """Add a health check."""
+
         mgmt_router = APIRouter(prefix=self.app.state.router_prefix)
 
-        @mgmt_router.get("/_mgmt/ping")
         async def ping():
-            """Liveliness/readiness probe."""
+            """Liveliness probe."""
             return {"message": "PONG"}
 
+        mgmt_router.add_api_route(
+            name="Ping",
+            path="/_mgmt/ping",
+            response_model=Dict,
+            responses={
+                200: {
+                    "content": {
+                        MimeTypes.json.value: {},
+                    },
+                },
+            },
+            response_class=self.response_class,
+            methods=["GET"],
+            endpoint=ping,
+        )
+
+        mgmt_router.add_api_route(
+            name="Health",
+            path="/_mgmt/health",
+            response_model=Dict,
+            responses={
+                200: {
+                    "content": {
+                        MimeTypes.json.value: {},
+                    },
+                },
+            },
+            response_class=self.response_class,
+            methods=["GET"],
+            endpoint=self.health_check,
+        )
         self.app.include_router(mgmt_router, tags=["Liveliness/Readiness"])
 
     def add_route_dependencies(
-        self, scopes: List[Scope], dependencies=List[Depends]
+        self, scopes: List[Scope], dependencies: List[Depends]
     ) -> None:
         """Add custom dependencies to routes.
 
         Args:
-            scopes: list of scopes. Each scope should be a dict with a `path`
-                and `method` property.
-            dependencies: list of [FastAPI
-                dependencies](https://fastapi.tiangolo.com/tutorial/dependencies/)
+            scopes: list of Scope.
+                Each scope should be a dict with a `path` and `method` property.
+            dependencies: list of fastapi.Depends
+                [FastAPI dependencies](https://fastapi.tiangolo.com/tutorial/dependencies/)
                 to apply to each scope.
 
-        Returns:
-            None
         """
         return add_route_dependencies(self.app.router.routes, scopes, dependencies)
 
-    def __attrs_post_init__(self):
+    def __attrs_post_init__(self) -> None:
         """Post-init hook.
 
         Responsible for setting up the application upon instantiation of the class.
 
-        Returns:
-            None
         """
         # inject settings
         self.client.extensions = self.extensions
@@ -463,3 +468,6 @@ class StacApi:
         # customize route dependencies
         for scopes, dependencies in self.route_dependencies:
             self.add_route_dependencies(scopes=scopes, dependencies=dependencies)
+
+        if self.app.state.settings.enable_direct_response:
+            add_direct_response(self.app)