+++ stac_fastapi/types/stac_fastapi/types/core.py
@@ -653,7 +653,7 @@ class BaseCoreClient(LandingPageMixin, abc.ABC):
         bbox: Optional[BBox] = None,
         datetime: Optional[str] = None,
         limit: int = 10,
-        token: str = None,
+        token: Optional[str] = None,
         **kwargs,
     ) -> stac.ItemCollection:
         """Get all items from a specific collection.
@@ -864,7 +864,7 @@ class AsyncBaseCoreClient(LandingPageMixin, abc.ABC):
         bbox: Optional[BBox] = None,
         datetime: Optional[str] = None,
         limit: int = 10,
-        token: str = None,
+        token: Optional[str] = None,
         **kwargs,
     ) -> stac.ItemCollection:
         """Get all items from a specific collection.